refactor(value): single portable GC-safe memsetValues, drop the assembly
The assembly filled []Value with raw vector stores, bypassing GC write
barriers on the pointer field of Value — a latent use-after-free hazard.
Replace all build variants (assembly, archsimd, purego) with one
portable doubling-copy implementation in value.go: typedslicecopy
performs a bulk write barrier pass followed by a memmove, making the
fill GC-safe at memmove speed. Move the tests to value_internal_test.go
so every build exercises them.
