--- values_purego.go
@@ -1,9 +1,0 @@
-//go:build purego || !amd64
-
-package parquet
-
-func memsetValues(values []Value, model Value) {
-	for i := range values {
-		values[i] = model
-	}
-}
